Integrating chat and uploading into frontend
