+++ embedding_service.py
@@ -40,13 +40,14 @@ def create_embeddings():
     if not success:
         return "Could not chunk elements.", 400
 
-    success, data = get_data_one('Files', {'_id': ObjectId(file_id)}, {'chunk_ids': 1, 'file_name': 1})
+    success, data = get_data_one('Files', {'_id': ObjectId(file_id)}, {'chunk_ids': 1, 'file_name': 1, 'gcs_link': 1})
 
     if not success:
         return "Could not retrieve file", 400
     
     chunks = data['chunk_ids']
     file_name = data['file_name']
+    gcs_link = data['gcs_link']
 
     for chunk_id in chunks:
         text_to_embed = ""
@@ -59,7 +60,7 @@ def create_embeddings():
         if not success:
             return "Error"
         
-    success, error_message = update_one('Users', {'name': "Dummy", 'courses.course': course_code}, {'$push': {'courses.$.notes': {'_id': ObjectId(file_id), 'file_name': file_name}}})
+    success, error_message = update_one('Users', {'name': "Dummy", 'courses.course': course_code}, {'$push': {'courses.$.notes': {'_id': ObjectId(file_id), 'file_name': file_name, 'gcs_link': gcs_link}}})
 
     if not success:
         return error_message, 400
