+++ skulespark-next-js/pages/course_list.tsx
@@ -1,18 +1,18 @@
 import React, { useState, useEffect } from 'react';
+import { Note } from './course_note';
 
 interface CourseListProps {
-  onSelectCourse: (courseId: string) => void;
-  courses: Course[];
+  onSelectCourse: (course: Course) => void;
 }
 
 export interface Course {
   course: string;
-  notes: any[];
+  notes: Note[];
 }
 
-export const fetchCourses = async () => {
+export const fetchCourses = async (): Promise<Course[]> => {
   try {
-    const response = await fetch('http://127.0.0.1:5000/get_courses/69420', {
+    const response = await fetch('http://127.0.0.1:5000/file_structure/69420', {
       method: 'GET',
     });
 
@@ -24,33 +24,29 @@ export const fetchCourses = async () => {
     return data; // Assuming data is an array of courses directly
   } catch (error) {
     console.error('Error fetching courses:', error);
-    return null;
+    return [];
   }
 };
 
-const CourseList: React.FC<CourseListProps> = ({ onSelectCourse, courses: initialCourses }) => {
-  const [courses, setCourses] = useState<Course[]>(initialCourses);
+const CourseList: React.FC<CourseListProps> = ({ onSelectCourse }) => {
+  const [courses, setCourses] = useState<Course[]>([]);
 
   useEffect(() => {
     const fetchData = async () => {
       const data = await fetchCourses();
-      if (data) {
-        setCourses(data);
-      }
+      setCourses(data);
     };
 
     fetchData();
-  }, []); // Empty dependency array to run the effect only once
+  }, []);
 
   return (
     <div>
       <h2>Course List</h2>
       <ul>
-        {courses.map(course => (
+        {courses.map((course) => (
           <li key={course.course}>
-            <button onClick={() => onSelectCourse(course.course)}>
-              {course.course}
-            </button>
+            <button onClick={() => onSelectCourse(course)}>{course.course}</button>
           </li>
         ))}
       </ul>
