+++ skulespark-next-js/pages/skulespark_body.tsx
@@ -1,19 +1,38 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import './style/skulespark_body.css';
 import Notebook from './icons/notebook.png';
 import Image from 'next/image';
-import CourseList from './course_list';
-import CourseNotes from './course_note';
-import { Course } from './course_list';
-import { fetchCourses } from './course_list';
+import CourseList, { Course, fetchCourses } from './course_list';
+import CourseNotes, { Note } from './course_note';
+import MarkdownRenderer from '../pages/MarkdownRenderer';
+import AugmentedNote from '../pages/AugmentedNote';
 
-const SkuleSparkBody = () => {
+interface SkuleSparkBodyProps{
+  fileStructure: Course[];
+}
+
+const SkuleSparkBody = ({fileStructure}) => {
   const [showAdditionalColumns, setShowAdditionalColumns] = useState(false);
   const [selectedCourse, setSelectedCourse] = useState<Course | null>(null);
   const [showCreateCoursePopup, setShowCreateCoursePopup] = useState(false);
+  const [showUploadNotePopup, setShowUploadNotePopup] = useState(false);
   const [newCourseName, setNewCourseName] = useState('');
   const [courses, setCourses] = useState<Course[]>([]);
   const [courseListKey, setCourseListKey] = useState(0); // Add key state
+  const [selectedNote, setSelectedNote] = useState<Note | null>(null);
+  const [chatReady, setChatReady] = useState(false);
+  const [userInput, setUserInput] = useState('')
+  const [chatOutput, setChatOutput] = useState('')
+  const [sources, setSources] = useState([])
+  const [pdfFile, setPdfFile] = useState(null);
+  const [pdfURL, setPdfURL] = useState('');
+  const [ocrResult, setOCRResult] = useState([]);
+  const [notesElements, setNotesElements] = useState([])
+  const elementRefs = useRef<Array<HTMLDivElement | null>>([]);
+  const [noteType, setNoteType] = useState(null)
+  const [fileId, setFileId] = useState(null);
+  const [noteListKey, setNoteListKey] = useState(0);
+  
 
   const handleButtonClick = () => {
     setShowAdditionalColumns(!showAdditionalColumns);
@@ -24,6 +43,10 @@ const SkuleSparkBody = () => {
     setSelectedCourse(course);
   };
 
+  const handleSelectNote = (note: Note) => {
+    setSelectedNote(note);
+  };
+
   const handleCreateCourse = () => {
     setShowCreateCoursePopup(true);
   };
@@ -32,6 +55,50 @@ const SkuleSparkBody = () => {
     setShowCreateCoursePopup(false);
     setNewCourseName(''); // Reset the input field
   };
+
+  const handleUploadNote = () => {
+    setShowUploadNotePopup(true);
+    handleUpload();
+  }
+
+  const handleNoteTypeChange = (e) => {
+    setNoteType(e.target.value)
+  }
+
+  const handleUpload = async () => {
+    try {
+      if(noteType == null){
+        return // Should indicate some sort of alert to the user
+      }
+      
+      const formData = new FormData();
+      formData.append('pdf', pdfFile);
+      formData.append('note_type', noteType);
+      console.log(selectedCourse?.course)
+      formData.append('course', selectedCourse?.course);
+
+      const response = await fetch('http://127.0.0.1:5000/upload', {
+        method: 'POST',
+        body: formData,
+      });
+
+      const data = await response.json();
+      setPdfURL(data.gcs_pdf_url);
+      setOCRResult(data.ocr_result);
+      setFileId(data.file_id);
+
+      // Increment the key to force re-render of CourseNotes
+      setNoteListKey((prevKey) => prevKey + 1);
+
+    } catch (error) {
+      console.error('Error uploading PDF:', error);
+    }
+  };
+
+  const handleCancelUploadNote = () => {
+    setShowUploadNotePopup(false);
+  };
+  
 
   const handleConfirmCreateCourse = async () => {
     try {
@@ -55,6 +122,154 @@ const SkuleSparkBody = () => {
       }
     } catch (error) {
       console.error('Error during course creation:', error);
+    }
+  };
+
+  const updateChat = async () => {
+    // TODO: Add logic to generate new chat output based on the user input
+    // For now, let's echo the user input
+    try{
+      const response = await fetch('http://127.0.0.1:5000/chat-prompt', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ file_id: fileId, message: userInput }),
+      });
+      if (response.ok) {
+        const data = await response.json();
+        setChatOutput(`SkuleSpark: ${data.answer}`);
+        setSources(data.sources)
+        setUserInput('');
+      
+      }
+    }
+    catch (error) {
+      console.error('Error during chat:', error);
+    }
+    
+  };
+
+  const displayChat = () => {
+    if (!chatReady) {
+      return <div></div>;
+    }
+
+    return (
+      <div>
+        <h2>Chat with your recently updated Note:</h2>
+        <div style={{ display: 'flex' }}>
+          {/* Chat Input */}
+          <div style={{ flex: 1 }}>
+            <h2>Question</h2>
+            <input
+              type="text"
+              placeholder="Type your question..."
+              value={userInput}
+              onChange={(e) => setUserInput(e.target.value)}
+            />
+            <button onClick={updateChat}>Ask</button>
+          </div>
+
+          {/* Chat Output */}
+          <div style={{ flex: 1, textAlign: 'left' }}>
+            <h2>Answer</h2>
+            <MarkdownRenderer content={chatOutput} />
+            <h3>Sources: </h3>
+            {sources.map((source, index) => (
+              <div key={index}>
+                <p>{source}</p>
+              </div>
+            ))}
+          </div>
+        </div>
+      </div>
+    );
+  };
+
+  const handleFileChange = (e) => {
+    const selectedFile = e.target.files[0];
+    setPdfFile(selectedFile);
+  };
+
+  const displayOCRResult = () => {
+    if (ocrResult.length == 0) {
+      return <div></div>;
+    }
+
+    return (
+      <div>
+        <h2>OCR Results:</h2>
+        <div style={{ overflowY: 'auto', maxHeight: '600px'}}>
+        {ocrResult.map((result, index) => (
+          <div style={{border: '1px solid', padding: "5px 5px"}}
+            key={index}
+            contentEditable="true"
+            ref={(element) => (elementRefs.current[index] = element)}
+            data-id={result['id']}
+          >
+            <p>{result['text']}</p>
+          </div>
+        ))}
+        </div>
+        <button onClick={handleOCRConfirm}>Confirm OCR Results</button>
+      </div>
+    );
+  };
+
+  const displayAugmentedNotes = () => {
+    if (notesElements.length<=0){
+      return <div></div>;
+    }
+    
+    return <div className='augmented-notes-wrapper'> <AugmentedNote elements={notesElements} /> </div>
+  }
+
+  const getAugmentedNotes = async () =>{
+    try{
+      const response = await fetch(`http://127.0.0.1:5000/augmented-note/${fileId}`, {
+        method: 'GET',
+      });
+      if (response.ok) {
+        const data = await response.json();
+        console.log(data)
+        setNotesElements(data)
+      
+      }
+    }
+    catch (error) {
+      console.error('Error during fetching augmented notes:', error);
+    }
+  }
+  
+  const handleOCRConfirm = async () => {
+    // Send confirmation to embedding service
+    try{
+      const updatedElements = elementRefs.current.map((ref) => {
+        if (ref) {
+          return {
+            id: ref.dataset.id,
+            text: ref.innerText,
+          };
+        }
+      });
+
+      const response = await fetch('http://127.0.0.1:5000/confirm_results', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json', // Set the content type to JSON
+        },
+        body: JSON.stringify({file_id: fileId , confirmed_elements: updatedElements, course_code: "ECE334" /*TODO: Change this to dynamic*/}),
+      });
+      if (response.ok) {
+        // Request was successful
+        const data = await response.json();
+        setChatReady(true);
+        getAugmentedNotes();
+      }
+    }
+    catch (error) {
+      console.error('Error during OCR:', error);
     }
   };
 
@@ -89,17 +304,32 @@ const SkuleSparkBody = () => {
             </div>
           </div>
           <div className={`column ${showAdditionalColumns ? 'small' : ''}`}>
-            <CourseNotes selectedCourse={selectedCourse} />
+            <CourseNotes key={noteListKey} selectedCourse={selectedCourse} onSelectNote={handleSelectNote} fileStructure={fileStructure} />
+            <div>
+              {selectedCourse != null && (
+                <button onClick={handleUploadNote} className="uploadNoteButton">
+                  Upload Note
+                </button>
+              )}
+            </div>
           </div>
         </>
       )}
       <div className={`column column2 ${showAdditionalColumns ? 'small' : ''}`}>
-        Augmented Note would get shown here
+        {selectedNote ? (
+          <iframe
+            src={`https://docs.google.com/viewer?url=${selectedNote.gcs_link}&embedded=true`}
+            title="pdf-viewer"
+            width="100%"
+            height="100%"
+          />
+        ) : (
+          'No note selected. Select a note to view PDF!'
+        )}
       </div>
       <div className="column column3">
-        {selectedCourse
-          ? `Selected Course: ${selectedCourse.course}`
-          : 'No course selected. Select a course to view notes!'}
+        {/*{displayAugmentedNotes()}*/}
+        {displayChat()}
       </div>
 
       {/* Popup for creating a new course */}
@@ -113,6 +343,26 @@ const SkuleSparkBody = () => {
           />
           <button onClick={handleCancelCreateCourse}>Cancel</button>
           <button onClick={handleConfirmCreateCourse}>Confirm</button>
+        </div>
+      )}
+
+      {/* Popup for uploading a new note */}
+      {showUploadNotePopup && selectedCourse != null && (
+        <div className="popup">
+          <input type="file" onChange={handleFileChange} accept=".pdf" />
+          <button onClick={handleCancelUploadNote}>Cancel</button>
+          <button onClick={handleUploadNote}>Confirm</button>
+          <div>
+            <input type="radio" id="typed" name="note-type" value="typed" onChange={handleNoteTypeChange}/>
+            <label for="typed">Typed</label>
+
+            <input type="radio" id="handwritten" name="note-type" value="handwritten" onChange={handleNoteTypeChange}/>
+            <label for="handwritten">Handwritten</label>
+
+            <input type="radio" id="both" name="note-type" value="both" onChange={handleNoteTypeChange}/>
+            <label for="both">Typed + Handwritten</label>
+          </div>
+          {displayOCRResult()}
         </div>
       )}
     </div>
